+++ test/etc/workspace_1/main.ts
@@ -1,9 +1,9 @@
-import ImportOrganizer from './import-organizer';
-import ImportManager from './import-organizer/ImportManager';
-import iocSymbols, { ImportManagerProvider } from './ioc-symbols/index';
-import TypescriptHero from './typescript-hero';
+import iocSymbols from './ioc-symbols';
+import { Foo } from './server/indices/foobar';
 import winstonLogger, { Logger } from './utilities/Logger';
 
 function foo() {
   console.log(winstonLogger, Logger, iocSymbols);
 }
+
+Foo
